Added graphs, added age as a predictor
I added some exploratory graphs, pick the ones you think work best to show relationships in the data.
I added age as a predictor to the predictive model, it improved the accuracy by 6%.
Based on the graph of experience vs subscription I think we should remove it as a predictor since there isn't a relationship between those two variables.